extract and use type definitions
